corrected the comments, added the text of the tasks
