+++ HomeExercise.md
@@ -1,0 +1,107 @@
+# Задание TagsCloudContainer
+
+Сделай fork этого репозитория и работай в нем.
+
+Нужно сделать приложение генератора [облака слов](https://ru.wikipedia.org/wiki/%D0%9E%D0%B1%D0%BB%D0%B0%D0%BA%D0%BE_%D1%82%D0%B5%D0%B3%D0%BE%D0%B2) по произвольному тексту.
+[Примеры](https://www.google.ru/search?q=%D0%9E%D0%B1%D0%BB%D0%B0%D0%BA%D0%BE+%D1%81%D0%BB%D0%BE%D0%B2&tbm=isch).
+
+В облаке не должно быть повторяющихся слов, размер слова должен быть тем больше, чем чаще встречается слово, не должно быть "скучных" слов (предлогов, местоимений, ...).
+
+Используй алгоритм из блока про [TDD](https://github.com/kontur-courses/tdd).
+
+## Ограничения
+
+### Точки расширения
+
+В промышленном программировании при разработке далеко не всегда разумно создавать точки расширения "на будущее".
+Чаще руководствуются принципами YAGNI и KISS, поддерживая код как можно проще, а абстракции создаются и внедряются в код только в момент,
+когда понадобилось расширить функциональность.
+
+Тем не менее для учебных целей в этой задаче требуется заранее предусмотреть точки расширения для наиболее вероятных потенциальных изменений в твоем продукте.
+Результат твоей работы должен быть расширяем без модификации уже имеющегося кода (Принцип OCP).
+
+### Полиморфизм вместо условных операторов
+
+Запрещено использовать операторы if, switch, ?: и прочие условные операторы, если их можно заменить полиморфизмом.
+
+### Dependency injection
+
+Для сборки зависимостей используй любой DI Container, отличный от Microsoft.Extensions.DependencyInjection.
+Можно воспользоваться [Pure.DI](https://www.nuget.org/packages/Pure.DI/) или довольно распространенным di контейнером [Autofac](https://www.nuget.org/packages/Autofac/). В выборе di контейнера может помочь статья [IoC Container Benchmark - Performance comparison](https://www.palmmedia.de/Blog/2011/8/30/ioc-container-benchmark-performance-comparison).
+
+### Тесты
+
+Все нетривиальные части покрой модульными тестами.
+Добавь несколько более крупных тестов, проверяющих работу всей программы в сборе.
+
+## Функциональные требования и их возможные изменения
+
+Ниже описаны обязательные пункты и пункты на перспективу.
+Выполни обязательные требования, а потом выбери и реализуй несколько понравившихся пунктов на перспективу.
+
+Даже если требование из перспективы не выполнено, соответствующая точка расширения в вашем коде уже должна быть.
+- - -
+### Обязательные пункты:
+**Исходный текст**:
+*  Источником данных должен быть файл со словами по одному в строке.
+
+**Предобработка слов**:
+* На этапе предобработки, приведи все слова к нижнему регистру и исключи скучные слова
+
+**Формат результата:**
+* В качестве результата программа должна генерировать png-файл.
+* Должна быть возможность задать цвета, шрифт и размер изображения.
+
+**Алгоритм**
+* Сделай так, чтобы по одному тексту можно было сгенерировать несколько облаков тегов с помощью разных алгоритмов или одного алгоритма с разными настройками.
+
+**GUI или Console**:
+* Организуй код так, чтобы было легко сделать оба вида клиентов — и клиента командной строки, и GUI приложение с вводом параметров и интерактивным просмотром.
+* Реализуй одного клиента по выбору
+
+- - -
+
+### Пункты на перспективу:
+**Исходный текст**:
+* поддерживать разные форматы файлов (txt, doc, docx, ...)
+
+**Предобработка слов**:
+* дать возможность влиять на список скучных слов, которые не попадут в облако.
+* поддерживать ввод данных из литературного текста, с приведением слов в начальную форму.
+* дать возможность выбирать только определенные части речи (например, только существительные)
+
+**Формат результата**:
+* поддерживать разные форматы изображений.
+* поддерживать разные алгоритмы расцветки слов.
+
+**Алгоритм**:
+* реализовать несколько алгоритмов формирования облака тегов.
+
+**GUI или Console**:
+* реализуй второго клиента.
+
+- - -  
+
+## Критерии оценивания:
+0 баллов - в решении отсутствует метод Main для запуска приложения, падающие тесты, не выполнены обязательные требования требования
+
+1 балл - в решении есть какие-то проблемы, выполнены обязательные требования
+
+2 балла - хорошо выполнены обязательные требования
+
+3 балла - выполнены основыные пункты и реализованы минимум два пункта на перспективу из разных разделов.
+
+- - -  
+
+## Дополнительные ссылки
+
+* Библиотеки для разбора аргументов командной строки:
+
+    * https://github.com/commandlineparser/commandline
+    * https://github.com/natemcmaster/CommandLineUtils
+
+* Приведение слова к начальной форме и определение частей речи:
+    * Библиотека https://www.nuget.org/packages/WeCantSpell.Hunspell/
+    * Утилита командной строки MyStem https://tech.yandex.ru/mystem/doc/usage-examples-docpage/
+	
+
