+++ TagCloudConsoleApp/Interfaces/ISettingsProvider.cs
@@ -1,0 +1,8 @@
+using TagsCloudVisualization.Settings;
+
+namespace TagCloudConsoleApp.Interfaces;
+
+public interface ISettingsProvider
+{
+    public SettingsManager GetSettings();
+}
