--- TagsCloudVisualization/Client/IClient.cs
@@ -1,8 +1,0 @@
-using TagsCloudVisualization.Settings;
-
-namespace TagsCloudVisualization.Client;
-
-public interface IClient
-{
-    public SettingsManager GetSettings();
-}
