--- TagsCloudVisualization/CloudLayouter/ICloudLayouter.cs
@@ -1,9 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.CloudLayouter;
-
-public interface ICloudLayouter
-{
-    public List<Rectangle> Rectangles { get; }
-    public Rectangle PutNextRectangle(Size rectangleSize);
-}
