--- TagsCloudVisualization/Draw/IRectangleDraftsman.cs
@@ -1,9 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Draw;
-
-public interface IRectangleDraftsman
-{
-    public Bitmap Bitmap { get; }
-    public void CreateImage(IEnumerable<Rectangle> objects);
-}
