--- TagsCloudVisualization/Filter/BoringWordsTextFilter.cs
@@ -1,10 +1,0 @@
-namespace TagsCloudVisualization.Filter;
-
-public class BoringWordsTextFilter : ITextFilter
-{
-    public List<string> ApplyFilter(IEnumerable<string> text)
-    {
-        //TODO: implement boring words filter
-        throw new NotImplementedException();
-    }
-}
