--- TagsCloudVisualization/Filter/ITextFilter.cs
@@ -1,6 +1,0 @@
-namespace TagsCloudVisualization.Filter;
-
-public interface ITextFilter
-{
-    public List<string> ApplyFilter(IEnumerable<string> text);
-}
