--- TagsCloudVisualization/Filter/LowerCaseTextFilter.cs
@@ -1,9 +1,0 @@
-namespace TagsCloudVisualization.Filter;
-
-public class LowerCaseTextFilter : ITextFilter
-{
-    public List<string> ApplyFilter(IEnumerable<string> text)
-    {
-        return text.Select(word => word.ToLower()).ToList();
-    }
-}
