--- TagsCloudVisualization/Generator/IBitmapGenerator.cs
@@ -1,8 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Generator;
-
-public interface IBitmapGenerator
-{
-    public Bitmap GenerateBitmap(IEnumerable<TagWord> words);
-}
