--- TagsCloudVisualization/Generator/IPositionGenerator.cs
@@ -1,9 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Generator;
-
-public interface IPositionGenerator
-{
-    public Point Center { get; }
-    public Point GetNextPoint();
-}
