--- TagsCloudVisualization/Generator/IRectangleGenerator.cs
@@ -1,8 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Generator;
-
-public interface IRectangleGenerator
-{
-    public IEnumerable<Rectangle> GenerateRectangles(int countRectangles);
-}
