--- TagsCloudVisualization/Generator/RandomRectangleGenerator.cs
@@ -1,17 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Generator;
-
-public class RandomRectangleGenerator : IRectangleGenerator
-{
-    public IEnumerable<Rectangle> GenerateRectangles(int countRectangles)
-    {
-        var random = new Random();
-        
-        for (var i = 0; i < countRectangles; i++)
-        {
-            var size = new Size(random.Next(20,100), random.Next(10, 40));
-            yield return new Rectangle(new Point(), size);
-        }
-    }
-}
