--- TagsCloudVisualization/Generator/TagWord.cs
@@ -1,3 +1,0 @@
-namespace TagsCloudVisualization.Generator;
-
-public record TagWord(string Word, int FontSize);
