+++ TagsCloudVisualization/Interfaces/ICloudLayouter.cs
@@ -1,0 +1,9 @@
+using System.Drawing;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface ICloudLayouter
+{
+    public List<Rectangle> Rectangles { get; }
+    public Rectangle PutNextRectangle(Size rectangleSize);
+}
