+++ TagsCloudVisualization/Interfaces/IImageSaver.cs
@@ -1,0 +1,9 @@
+using System.Drawing;
+using TagsCloudVisualization.Settings;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface IImageSaver
+{
+    public string SaveImageToFile(Bitmap bitmap, SaveSettings settings);
+}
