+++ TagsCloudVisualization/Interfaces/IPositionGenerator.cs
@@ -1,0 +1,9 @@
+using System.Drawing;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface IPositionGenerator
+{
+    public Point Center { get; }
+    public Point GetNextPoint();
+}
