+++ TagsCloudVisualization/Interfaces/IRectangleDraftsman.cs
@@ -1,0 +1,9 @@
+using System.Drawing;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface IRectangleDraftsman
+{
+    public Bitmap Bitmap { get; }
+    public void CreateImage(IEnumerable<Rectangle> objects);
+}
