+++ TagsCloudVisualization/Interfaces/IRectangleGenerator.cs
@@ -1,0 +1,8 @@
+using System.Drawing;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface IRectangleGenerator
+{
+    public IEnumerable<Rectangle> GenerateRectangles(int countRectangles);
+}
