+++ TagsCloudVisualization/Interfaces/ITextFilter.cs
@@ -1,0 +1,6 @@
+namespace TagsCloudVisualization.Interfaces;
+
+public interface ITextFilter
+{
+    public List<string> ApplyFilter(IEnumerable<string> text);
+}
