+++ TagsCloudVisualization/Interfaces/ITextReader.cs
@@ -1,0 +1,6 @@
+namespace TagsCloudVisualization.Interfaces;
+
+public interface ITextReader
+{
+    public List<string> ReadText();
+}
