+++ TagsCloudVisualization/Models/Filters/BoringWordsTextFilter.cs
@@ -1,0 +1,12 @@
+using TagsCloudVisualization.Interfaces;
+
+namespace TagsCloudVisualization.Models.Filters;
+
+public class BoringWordsTextFilter : ITextFilter
+{
+    public List<string> ApplyFilter(IEnumerable<string> text)
+    {
+        //TODO: implement boring words filter
+        throw new NotImplementedException();
+    }
+}
