+++ TagsCloudVisualization/Models/Filters/LowerCaseTextFilter.cs
@@ -1,0 +1,11 @@
+using TagsCloudVisualization.Interfaces;
+
+namespace TagsCloudVisualization.Models.Filters;
+
+public class LowerCaseTextFilter : ITextFilter
+{
+    public List<string> ApplyFilter(IEnumerable<string> text)
+    {
+        return text.Select(word => word.ToLower()).ToList();
+    }
+}
