+++ TagsCloudVisualization/Models/Generators/TagCloudImageGenerator.cs
@@ -1,0 +1,41 @@
+using TagsCloudVisualization.Interfaces;
+using TagsCloudVisualization.Settings;
+
+namespace TagsCloudVisualization.Models.Generators;
+
+public class TagCloudImageGenerator(
+    IImageSaver saver,
+    SaveSettings settings,
+    ITextReader reader,
+    IBitmapGenerator bitmapGenerator,
+    IEnumerable<ITextFilter> filters)
+{
+    private const int MaxFontSize = 100;
+    private const int MinFontSize = 8;
+    private int maxWordCount;
+    private int minWordCount;
+
+    public void GenerateCloud()
+    {
+        var text = reader.ReadText();
+
+        var wordsFrequency = filters
+            .Aggregate(text, (word, filter) => filter.ApplyFilter(word))
+            .GroupBy(w => w)
+            .OrderByDescending(words => words.Count())
+            .ToDictionary(words => words.Key, words => words.Count());
+
+        minWordCount = wordsFrequency.Values.Min();
+        maxWordCount = wordsFrequency.Values.Max();
+
+        var words = wordsFrequency
+            .Select(w => new TagWord(w.Key, GetFontSize(w.Value)));
+
+        var bitmap = bitmapGenerator.GenerateBitmap(words);
+        saver.SaveImageToFile(bitmap, settings);
+    }
+
+    private int GetFontSize(int frequencyCount) =>
+        MinFontSize + (MaxFontSize - MinFontSize)
+        * (frequencyCount - minWordCount) / (maxWordCount - minWordCount);
+}
