+++ TagsCloudVisualization/Models/Generators/TagWord.cs
@@ -1,0 +1,3 @@
+namespace TagsCloudVisualization.Models.Generators;
+
+public record TagWord(string Word, int FontSize);
