+++ TagsCloudVisualization/Models/Visualizatiuons/RectangleVisualizatiuon.cs
@@ -1,0 +1,34 @@
+using System.Drawing;
+using TagsCloudVisualization.Interfaces;
+
+namespace TagsCloudVisualization.Models.Visualizatiuons;
+
+public class RectangleVisualizatiuon : IRectangleDraftsman
+{
+    public Bitmap Bitmap => new(bitmap);
+    private readonly Bitmap bitmap;
+    private readonly Size shiftToBitmapCenter;
+
+    public RectangleVisualizatiuon(int width, int height)
+    {
+        #pragma warning disable CA1416
+        bitmap = new(width, height);
+        shiftToBitmapCenter = new Size(bitmap.Width / 2, bitmap.Height / 2);
+        #pragma warning restore CA1416
+    }
+
+    public void CreateImage(IEnumerable<Rectangle> rectangles)
+    {
+        ArgumentNullException.ThrowIfNull(rectangles);
+        
+        #pragma warning disable CA1416
+        using var graphics = Graphics.FromImage(bitmap);
+        graphics.Clear(Color.White);
+        foreach (var r in rectangles)
+        {
+            var rectangle = new Rectangle(r.Location + shiftToBitmapCenter, r.Size);
+            graphics.DrawRectangle(new Pen(Color.BlueViolet), rectangle);
+        }
+        #pragma warning restore CA1416
+    }
+}
