--- TagsCloudVisualization/Reader/ITextReader.cs
@@ -1,6 +1,0 @@
-namespace TagsCloudVisualization.Reader;
-
-public interface ITextReader
-{
-    public List<string> ReadText();
-}
