--- TagsCloudVisualization/Saver/IImageSaver.cs
@@ -1,8 +1,0 @@
-using System.Drawing;
-
-namespace TagsCloudVisualization.Saver;
-
-public interface IImageSaver
-{
-    public string SaveImageToFile(Bitmap bitmap);
-}
