+++ TagsCloudVisualization/Settings/SpiralGeneratorSettings.cs
@@ -2,9 +2,4 @@ using System.Drawing;
 
 namespace TagsCloudVisualization.Settings;
 
-public class SpiralGeneratorSettings(double angleOffset, double spiralStep, Point center)
-{
-    public double AngleOffset { get; private set; } = angleOffset;
-    public double SpiralStep { get; private set; } = spiralStep;
-    public Point Center { get; private set; } = center;
-}
+public record SpiralGeneratorSettings(double AngleOffset, double SpiralStep, Point Center);
