+++ TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -12,6 +12,7 @@
 
     <ItemGroup>
       <Folder Include="Images\" />
+      <Folder Include="Models\" />
     </ItemGroup>
 
 </Project>
